+++ test/tables/fvar.spec.mjs
@@ -1,8 +1,8 @@
 import assert from 'assert';
 import { hex, unhex } from '../testutil.mjs';
-import fvar from '../../src/tables/fvar.mjs';
 import { Font, parse } from '../../src/opentype.mjs';
 import { readFileSync } from 'fs';
+import { makeFvarTable, parseFvarTable } from '../../src/fn/index.mjs';
 const loadSync = (url, opt) => parse(readFileSync(url), opt);
 
 describe('tables/fvar.mjs', function() {
@@ -62,7 +62,7 @@ describe('tables/fvar.mjs', function() {
     };
 
     it('can parse a font variations table', function() {
-        assert.deepEqual(table, fvar.parse(unhex(data), 0, names));
+        assert.deepEqual(table, parseFvarTable(unhex(data), 0, names));
     });
 
     it('parses nameIDs 2 and 17 and postScriptNameID 6 correctly', function() {
@@ -106,12 +106,12 @@ describe('tables/fvar.mjs', function() {
                 257: {en: 'Weight', ja: 'ウエイト'}
             }
         };
-        assert.deepEqual(data, hex(fvar.make(table, names).encode()));
+        assert.deepEqual(data, hex(makeFvarTable(table, names).encode()));
     });
 
     it('writes postScriptNameID optionally', function() {
         let parsedFont = parse(testFont.toArrayBuffer());
-        let makeTable = fvar.make(parsedFont.tables.fvar, parsedFont.names);
+        let makeTable = makeFvarTable(parsedFont.tables.fvar, parsedFont.names);
         assert.equal(parsedFont.tables.fvar.instances[0].postScriptNameID, 6);
         assert.equal(parsedFont.tables.fvar.instances[0].postScriptName.en, 'VARTestVF-Regular');
 
@@ -121,7 +121,7 @@ describe('tables/fvar.mjs', function() {
             parsedFont.tables.fvar.instances.map(i => { i.postScriptNameID = undefined; return i; });
 
         parsedFont = parse(parsedFont.toArrayBuffer());
-        makeTable = fvar.make(parsedFont.tables.fvar, parsedFont.names);
+        makeTable = makeFvarTable(parsedFont.tables.fvar, parsedFont.names);
         
         assert.equal(makeTable.instanceSize, 8);
